+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.3.0-rc.3",
-  genVersion: "2.799.0",
-  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.3 2.799.0 2.0 @orq-ai/node",
+  sdkVersion: "4.3.0-rc.5",
+  genVersion: "2.801.0",
+  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.5 2.801.0 2.0 @orq-ai/node",
 } as const;