+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-26T12:36:54.881Z",
+    "2026-01-27T12:37:49.956Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {