+++ packages/orq-rc/RELEASES.md
@@ -6199,3 +6199,13 @@ Based on:
 - [typescript v4.3.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.3 - packages/orq-rc
+
+## 2026-01-27 12:40:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.698.0 (2.801.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.5 - packages/orq-rc